Move XMPP transport to separate file
